+++ Headers/DebugServer2/Core/SoftwareBreakpointManager.h
@@ -11,24 +11,48 @@
 #ifndef __DebugServer2_SoftwareBreakpointManager_h
 #define __DebugServer2_SoftwareBreakpointManager_h
 
-#include "DebugServer2/Base.h"
-
-#if defined(ARCH_ARM) || defined(ARCH_ARM64)
-#include "DebugServer2/Architecture/ARM/SoftwareBreakpointManager.h"
-#elif defined(ARCH_X86) || defined(ARCH_X86_64)
-#include "DebugServer2/Architecture/X86/SoftwareBreakpointManager.h"
-#else
-#error "Architecture not supported."
-#endif
+#include "DebugServer2/Core/BreakpointManager.h"
 
 namespace ds2 {
+class SoftwareBreakpointManager : public BreakpointManager {
+private:
+  std::map<uint64_t, ByteVector> _insns;
+
+public:
+  SoftwareBreakpointManager(Target::ProcessBase *process);
+  ~SoftwareBreakpointManager() override;
+
+public:
+  virtual void clear() override;
+
+public:
+  virtual int hit(Target::Thread *thread, Site &site) override;
+
+protected:
+  virtual void getOpcode(uint32_t type, ByteVector &opcode) const;
+
+protected:
+  virtual ErrorCode enableLocation(Site const &site) override;
+  virtual ErrorCode disableLocation(Site const &site) override;
+
+protected:
+  ErrorCode isValid(Address const &address, size_t size,
+                    Mode mode) const override;
+
 #if defined(ARCH_ARM) || defined(ARCH_ARM64)
-using Architecture::ARM::SoftwareBreakpointManager;
-#elif defined(ARCH_X86) || defined(ARCH_X86_64)
-using Architecture::X86::SoftwareBreakpointManager;
-#else
-#error "Architecture not supported."
+public:
+  virtual void
+  enumerate(std::function<void(Site const &)> const &cb) const override;
+
+public:
+  virtual ErrorCode add(Address const &address, Type type, size_t size,
+                        Mode mode) override;
+  virtual ErrorCode remove(Address const &address) override;
+
+public:
+  virtual bool has(Address const &address) const override;
 #endif
+};
 }
 
 #endif // !__DebugServer2_SoftwareBreakpointManager_h
